# internal/clientgen/testdata/goapp/expected_baseauth_openapi.json
{
  "components": {
    "responses": {
      "APIError": {
        "content": {
          "application/json": {
            "schema": {
              "externalDocs": {
                "url": "https://pkg.go.dev/encore.dev/beta/errs#Error"
              },
              "properties": {
                "code": {
                  "description": "Error code",
                  "example": "not_found",
                  "externalDocs": {
                    "url": "https://pkg.go.dev/encore.dev/beta/errs#ErrCode"
                  },
                  "type": "string"
                },
                "details": {
                  "description": "Error details",
                  "type": "object"
                },
                "message": {
                  "description": "Error message",
                  "type": "string"
                }
              },
              "title": "APIError",
              "type": "object"
            }
          }
        },
        "description": "Error response"
      }
    }
  },
  "info": {
    "description": "Generated by encore",
    "title": "API for app",
    "version": "1",
    "x-logo": {
      "altText": "Encore logo",
      "backgroundColor": "#EEEEE1",
      "url": "https://encore.dev/assets/branding/logo/logo-black.png"
    }
  },
  "openapi": "3.0.0",
  "paths": {
    "/svc.DummyAPI": {
      "post": {
        "operationId": "POST:svc.DummyAPI",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "Message": {
                    "type": "string"
                  }
                },
                "required": [
                  "Message"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        },
        "summary": "DummyAPI is a dummy endpoint.\n"
      }
    },
    "/svc.Private": {
      "post": {
        "operationId": "POST:svc.Private",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "Message": {
                    "type": "string"
                  }
                },
                "required": [
                  "Message"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        },
        "summary": "Private is a basic auth endpoint.\n"
      }
    }
  },
  "servers": [
    {
      "description": "Encore local dev environment",
      "url": "http://localhost:4000"
    }
  ]
}